# perf(VDataTable): reduce reactivity overhead for readonly items arrays